+++ src/main/java/com/codecool/WareStoreProject/model/mapper/WarehouseMapper.java
@@ -12,7 +12,7 @@ public class WarehouseMapper implements RowMapper<Warehouse> {
     @Override
     public Warehouse mapRow(ResultSet rs, int rowNum) throws SQLException {
         Warehouse warehouse = new Warehouse();
-        warehouse.setId(rs.getInt("id"));
+        warehouse.setId(rs.getLong("id"));
         warehouse.setName(rs.getString("name"));
         warehouse.setAddress(rs.getString("address"));
         warehouse.setStorageSpace(rs.getInt("storage_space"));
